requirements.txt
telethon==1.39.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0